product auth

+++ tests/test_product.py
@@ -1,16 +1,17 @@
 from fastapi.testclient import TestClient
 from fastapi import status
 
-def test_read_products(client: TestClient):
+def test_read_products(client: TestClient, admin_token_headers: dict):
     client.post(
         "/product/",
         json={
             "title": "Star Wars",
             "description": "Movie about space wars",
         },
+        headers=admin_token_headers
     )
 
-    response = client.get("/product/?limit=1")
+    response = client.get("/product/?limit=1", headers=admin_token_headers)
     data = response.json()
     
     assert data["count"] == 1      
@@ -22,17 +23,18 @@ def test_read_products(client: TestClient):
     assert result["title"] == "Star Wars"
     assert result["description"] == "Movie about space wars"
 
-def test_read_product_by_id(client: TestClient):
+def test_read_product_by_id(client: TestClient, admin_token_headers: dict):
     create_response = client.post(
         "/product/",
         json={
             "title": "Star Wars",
             "description": "Movie about space wars",
         },
+        headers=admin_token_headers
     )
     product_id = create_response.json()["id"]
 
-    response = client.get(f"/product/{product_id}")
+    response = client.get(f"/product/{product_id}", headers=admin_token_headers)
     
     assert response.status_code == status.HTTP_200_OK
     result = response.json()
@@ -40,13 +42,14 @@ def test_read_product_by_id(client: TestClient):
     assert result["title"] == "Star Wars"
     assert result["description"] == "Movie about space wars"
 
-def test_create_product(client: TestClient):
+def test_create_product(client: TestClient, admin_token_headers: dict):
     response = client.post(
         "/product/",
         json={
             "title": "Star Wars",
             "description": "Movie about space wars",
         },
+        headers=admin_token_headers
     )
 
     assert response.status_code == status.HTTP_200_OK
@@ -55,13 +58,14 @@ def test_create_product(client: TestClient):
     assert result["title"] == "Star Wars"
     assert result["description"] == "Movie about space wars"
 
-def test_update_product(client: TestClient):
+def test_update_product(client: TestClient, admin_token_headers: dict):
     create_response = client.post(
         "/product/",
         json={
             "title": "Star Wars",
             "description": "Movie about space wars",
         },
+        headers=admin_token_headers
     )
     product_id = create_response.json()["id"]
 
@@ -71,6 +75,7 @@ def test_update_product(client: TestClient):
             "title": "Star Wars 2",
             "description": "Movie about space wars 2",
         },
+        headers=admin_token_headers
     )
 
     assert response.status_code == status.HTTP_200_OK
@@ -79,15 +84,16 @@ def test_update_product(client: TestClient):
     assert result["title"] == "Star Wars 2"
     assert result["description"] == "Movie about space wars 2"
 
-def test_delete_product(client: TestClient):
+def test_delete_product(client: TestClient, admin_token_headers: dict):
     create_response = client.post(
         "/product/",
         json={
             "title": "Star Wars",
             "description": "Movie about space wars",
         },
+        headers=admin_token_headers
     )
     product_id = create_response.json()["id"]
 
-    response = client.delete(f"/product/{product_id}")
+    response = client.delete(f"/product/{product_id}", headers=admin_token_headers)
     assert response.status_code == status.HTTP_200_OK
